+++ internal/provider/vaultradar/resource_radar_integration_slack_connection_test.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package vaultradar_test
 
 import (
@@ -6,6 +9,7 @@ import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/plancheck"
 	"github.com/hashicorp/terraform-provider-hcp/internal/provider/acctest"
 )
 
@@ -16,10 +20,15 @@ func TestRadarIntegrationSlackConnection(t *testing.T) {
 	// Requires the following environment variables to be set:
 	projectID := os.Getenv("HCP_PROJECT_ID")
 	token := os.Getenv("RADAR_INTEGRATION_SLACK_TOKEN")
+	updateToken := os.Getenv("RADAR_INTEGRATION_SLACK_TOKEN_2")
 
-	if projectID == "" || token == "" {
-		t.Skip("HCP_PROJECT_ID and RADAR_INTEGRATION_SLACK_TOKEN must be set for acceptance tests")
+	if projectID == "" || token == "" || updateToken == "" {
+		t.Skip("HCP_PROJECT_ID, RADAR_INTEGRATION_SLACK_TOKEN and RADAR_INTEGRATION_SLACK_TOKEN_2 " +
+			"must be set for acceptance tests")
 	}
+
+	name := "AC Test of Creating Slack Connect from TF"
+	updatedName := "AC Test of Updating Slack Connect from TF"
 
 	resource.Test(t, resource.TestCase{
 		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
@@ -29,16 +38,58 @@ func TestRadarIntegrationSlackConnection(t *testing.T) {
 				Config: fmt.Sprintf(`
 					resource "hcp_vault_radar_integration_slack_connection" "example" {
 						project_id = %q
-						name = "AC Test of Slack Connect from TF"
+						name = %q
 						token = %q	
 					}
-				`, projectID, token),
+				`, projectID, name, token),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("hcp_vault_radar_integration_slack_connection.example", "project_id", projectID),
+					resource.TestCheckResourceAttr("hcp_vault_radar_integration_slack_connection.example", "name", name),
 					resource.TestCheckResourceAttrSet("hcp_vault_radar_integration_slack_connection.example", "id"),
 				),
 			},
-			// UPDATE not supported at this time.
+			// Update name.
+			{
+				Config: fmt.Sprintf(`
+					resource "hcp_vault_radar_integration_slack_connection" "example" {
+						project_id = %q
+						name = %q
+						token = %q	
+					}
+				`, projectID, updatedName, token),
+				ConfigPlanChecks: resource.ConfigPlanChecks{
+					PreApply: []plancheck.PlanCheck{
+						plancheck.ExpectResourceAction("hcp_vault_radar_integration_slack_connection.example", plancheck.ResourceActionUpdate),
+					},
+				},
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("hcp_vault_radar_integration_slack_connection.example", "name", updatedName),
+				),
+			},
+			// Update token. This effectively cause an update in the auth_key of the connection.
+			{
+				Config: fmt.Sprintf(`
+					resource "hcp_vault_radar_integration_slack_connection" "example" {
+						project_id = %q
+						name = %q
+						token = %q	
+					}
+				`, projectID, updatedName, updateToken),
+				ConfigPlanChecks: resource.ConfigPlanChecks{
+					PreApply: []plancheck.PlanCheck{
+						plancheck.ExpectResourceAction("hcp_vault_radar_integration_slack_connection.example", plancheck.ResourceActionUpdate),
+					},
+				},
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttrWith("hcp_vault_radar_integration_slack_connection.example", "token", func(value string) error {
+						if value != updateToken {
+							// Avoid outputting the token in the error message.
+							return fmt.Errorf("expected token to be updated")
+						}
+						return nil
+					}),
+				),
+			},
 			// DELETE happens automatically.
 		},
 	})