+++ internal/provider/vaultradar/resource_radar_source_github_enterprise_test.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package vaultradar_test
 
 import (
@@ -6,6 +9,7 @@ import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/plancheck"
 	"github.com/hashicorp/terraform-provider-hcp/internal/provider/acctest"
 )
 
@@ -18,9 +22,11 @@ func TestRadarSourceGitHubEnterprise(t *testing.T) {
 	githubOrganization := os.Getenv("RADAR_GITHUB_ENTERPRISE_ORGANIZATION")
 	domainName := os.Getenv("RADAR_GITHUB_ENTERPRISE_DOMAIN")
 	token := os.Getenv("RADAR_GITHUB_ENTERPRISE_TOKEN")
+	updateToken := os.Getenv("RADAR_GITHUB_ENTERPRISE_TOKEN_2")
 
-	if projectID == "" || githubOrganization == "" || domainName == "" || token == "" {
-		t.Skip("HCP_PROJECT_ID, RADAR_GITHUB_ENTERPRISE_ORGANIZATION, RADAR_GITHUB_ENTERPRISE_DOMAIN, and RADAR_GITHUB_ENTERPRISE_TOKEN must be set for acceptance tests")
+	if projectID == "" || githubOrganization == "" || domainName == "" || token == "" || updateToken == "" {
+		t.Skip("HCP_PROJECT_ID, RADAR_GITHUB_ENTERPRISE_ORGANIZATION, RADAR_GITHUB_ENTERPRISE_DOMAIN, " +
+			"RADAR_GITHUB_ENTERPRISE_TOKEN and RADAR_GITHUB_ENTERPRISE_TOKEN_2 must be set for acceptance tests")
 	}
 
 	resource.Test(t, resource.TestCase{
@@ -43,7 +49,31 @@ func TestRadarSourceGitHubEnterprise(t *testing.T) {
 					resource.TestCheckResourceAttrSet("hcp_vault_radar_source_github_enterprise.example", "id"),
 				),
 			},
-			// UPDATE not supported at this time.
+			// Update Token
+			{
+				Config: fmt.Sprintf(`
+					resource "hcp_vault_radar_source_github_enterprise" "example" {
+						project_id = %q
+						github_organization = %q
+						domain_name = %q
+						token = %q
+					}				
+				`, projectID, githubOrganization, domainName, updateToken),
+				ConfigPlanChecks: resource.ConfigPlanChecks{
+					PreApply: []plancheck.PlanCheck{
+						plancheck.ExpectResourceAction("hcp_vault_radar_source_github_enterprise.example", plancheck.ResourceActionUpdate),
+					},
+				},
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttrWith("hcp_vault_radar_source_github_enterprise.example", "token", func(value string) error {
+						if value != updateToken {
+							// Avoid outputting the token in the error message.
+							return fmt.Errorf("expected token to be updated")
+						}
+						return nil
+					}),
+				),
+			},
 			// DELETE happens automatically.
 		},
 	})