+++ internal/provider/vaultsecrets/resource_vault_secrets_integration_twilio_test.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package vaultsecrets_test
 
 import (
@@ -5,8 +8,8 @@ import (
 	"os"
 	"testing"
 
-	"github.com/hashicorp/hcp-sdk-go/clients/cloud-vault-secrets/preview/2023-11-28/client/secret_service"
-	secretmodels "github.com/hashicorp/hcp-sdk-go/clients/cloud-vault-secrets/preview/2023-11-28/models"
+	"github.com/hashicorp/hcp-sdk-go/clients/cloud-vault-secrets/stable/2023-11-28/client/secret_service"
+	secretmodels "github.com/hashicorp/hcp-sdk-go/clients/cloud-vault-secrets/stable/2023-11-28/models"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	"github.com/hashicorp/terraform-provider-hcp/internal/clients"
@@ -51,7 +54,7 @@ func TestAccVaultSecretsResourceIntegrationTwilio(t *testing.T) {
 				PreConfig: func() {
 					t.Helper()
 					client := acctest.HCPClients(t)
-					_, err := client.VaultSecretsPreview.DeleteTwilioIntegration(&secret_service.DeleteTwilioIntegrationParams{
+					_, err := client.VaultSecrets.DeleteTwilioIntegration(&secret_service.DeleteTwilioIntegrationParams{
 						Name:           integrationName2,
 						OrganizationID: client.Config.OrganizationID,
 						ProjectID:      client.Config.ProjectID,
@@ -72,7 +75,7 @@ func TestAccVaultSecretsResourceIntegrationTwilio(t *testing.T) {
 				PreConfig: func() {
 					t.Helper()
 					client := acctest.HCPClients(t)
-					_, err := client.VaultSecretsPreview.CreateTwilioIntegration(&secret_service.CreateTwilioIntegrationParams{
+					_, err := client.VaultSecrets.CreateTwilioIntegration(&secret_service.CreateTwilioIntegrationParams{
 						Body: &secretmodels.SecretServiceCreateTwilioIntegrationBody{
 							Capabilities: []*secretmodels.Secrets20231128Capability{secretmodels.Secrets20231128CapabilityROTATION.Pointer()},
 							StaticCredentialDetails: &secretmodels.Secrets20231128TwilioStaticCredentialsRequest{
@@ -143,7 +146,7 @@ func twilioIntegrationExists(t *testing.T, name string) bool {
 
 	client := acctest.HCPClients(t)
 
-	response, err := client.VaultSecretsPreview.GetTwilioIntegration(
+	response, err := client.VaultSecrets.GetTwilioIntegration(
 		secret_service.NewGetTwilioIntegrationParamsWithContext(ctx).
 			WithOrganizationID(client.Config.OrganizationID).
 			WithProjectID(client.Config.ProjectID).