+++ internal/provider/vaultsecrets/rotating_secret_twilio.go
@@ -6,8 +6,8 @@ package vaultsecrets
 import (
 	"context"
 
-	"github.com/hashicorp/hcp-sdk-go/clients/cloud-vault-secrets/preview/2023-11-28/client/secret_service"
-	secretmodels "github.com/hashicorp/hcp-sdk-go/clients/cloud-vault-secrets/preview/2023-11-28/models"
+	"github.com/hashicorp/hcp-sdk-go/clients/cloud-vault-secrets/stable/2023-11-28/client/secret_service"
+	secretmodels "github.com/hashicorp/hcp-sdk-go/clients/cloud-vault-secrets/stable/2023-11-28/models"
 	"github.com/hashicorp/terraform-provider-hcp/internal/clients"
 )
 
@@ -21,7 +21,7 @@ func (s *twilioRotatingSecret) read(ctx context.Context, client secret_service.C
 			WithOrganizationID(secret.OrganizationID.ValueString()).
 			WithProjectID(secret.ProjectID.ValueString()).
 			WithAppName(secret.AppName.ValueString()).
-			WithSecretName(secret.Name.ValueString()), nil)
+			WithName(secret.Name.ValueString()), nil)
 	if err != nil && !clients.IsResponseCodeNotFound(err) {
 		return nil, err
 	}
@@ -40,7 +40,7 @@ func (s *twilioRotatingSecret) create(ctx context.Context, client secret_service
 			WithBody(&secretmodels.SecretServiceCreateTwilioRotatingSecretBody{
 				RotationPolicyName: secret.RotationPolicyName.ValueString(),
 				IntegrationName:    secret.IntegrationName.ValueString(),
-				SecretName:         secret.Name.ValueString(),
+				Name:               secret.Name.ValueString(),
 			}),
 		nil)
 	if err != nil {